# openapi.json
{
  "components": {
    "schemas": {
...
      "ApiError": {
        "description": "One envelope for every error response: `{ \"code\", \"message\" }`.",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
...
      "PlaceTradeRequest": {
        "properties": {
          "action": {
            "$ref": "#/components/schemas/TradeActionDto"
          },
          "amount_micro": {
            "format": "int64",
            "type": "integer"
          },
          "expected_config_version": {
            "description": "The preview's `config_version` echo (D36: required).",
            "format": "int64",
            "type": "integer"
          },
          "idempotency_key": {
            "type": "string"
          },
          "market_ref": {
            "type": "string"
          },
          "pending_action_id": {
            "format": "uuid",
            "type": [
              "string",
              "null"
            ]
          },
          "run_id": {
            "format": "uuid",
            "type": [
              "string",
              "null"
            ]
          },
          "side": {
            "$ref": "#/components/schemas/SideDto"
          },
          "user_id": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "market_ref",
          "side",
          "action",
          "amount_micro",
          "idempotency_key",
          "expected_config_version"
        ],
        "type": "object"
      },
...
      "SideDto": {
        "enum": [
          "yes",
          "no"
        ],
        "type": "string"
      },
...
      "TradeActionDto": {
        "enum": [
          "buy",
          "sell"
        ],
        "type": "string"
      },
...
      "TradeReceiptDto": {
        "properties": {
          "action": {
            "$ref": "#/components/schemas/TradeActionDto"
          },
          "avg_price_micro": {
            "format": "int64",
            "type": "integer"
          },
          "fee_micro": {
            "format": "int64",
            "type": "integer"
          },
          "gross_micro": {
            "format": "int64",
            "type": "integer"
          },
          "ledger_txn": {
            "format": "uuid",
            "type": "string"
          },
          "replayed": {
            "type": "boolean"
          },
          "shares_micro": {
            "format": "int64",
            "type": "integer"
          },
          "side": {
            "$ref": "#/components/schemas/SideDto"
          },
          "trade_id": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "trade_id",
          "ledger_txn",
          "side",
          "action",
          "shares_micro",
          "gross_micro",
          "fee_micro",
          "avg_price_micro",
          "replayed"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "",
    "license": {
      "name": ""
    },
    "title": "adapters",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/trades": {
      "post": {
        "operationId": "place_trade",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlaceTradeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TradeReceiptDto"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": ""
          },
          "423": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "trades"
        ]
      }
    },
...
  },
...
}